+++ CHANGES.md
@@ -1,4 +1,7 @@
-- Fix errors caused by missing start locations in dragonflight season 1 dungeons
+- Add missing start locations for dragonflight S1 dungeons
+- Switch to enemy forces mode for dungeons with no known start locations
 - Fix not fading some frames when fade is enabled
 - Make fading progress smoothly between values
 - Add option to hide window completely while in combat
+- Remember current route when reloading/relogging or leaving the dungeon
+- Ignore mobs summoned during combat in route estimation